+++ examples/jest/jest.config.js
@@ -1,7 +1,17 @@
 module.exports = {
     "reporters": [
         "default",
-        ["<rootDir>/node_modules/kelonio/out/plugin/jestReporter", {keepStateAtStart: false, keepStateAtEnd: false, printReportAtEnd: true}]
+        [
+            "<rootDir>/node_modules/kelonio/out/plugin/jestReporter",
+            {
+                keepStateAtStart: false,
+                keepStateAtEnd: false,
+                printReportAtEnd: true,
+                extraReports: [
+                    { module: `${__dirname}/kelonioExtra.js`, callback: "extraReport" },
+                ],
+            },
+        ],
     ],
     "transform": {
         "^.+\\.tsx?$": "ts-jest",