+++ examples/karma/karma.conf.js
@@ -25,6 +25,9 @@ module.exports = config => {
         kelonioReporter: {
             inferBrowsers: true,
             printReportAtEnd: true,
+            extraReports: [
+                { module: `${__dirname}/kelonioExtra.js`, callback: "extraReport" },
+            ],
         },
     });
 };